Refactor EventTableDataModel out to a separate class
